+++ rust/tool/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.3.0"
 edition = "2021"
 
 [[bin]]
-name = "lzbt"
-path = "src/main.rs"
+name = "lzbt-systemd"
+path = "src/systemd-main.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
