+++ rust/tool/src/common/generation.rs
