+++ rust/tool/src/common/signature.rs
