+++ rust/tool/src/common/utils.rs
