+++ rust/tool/src/systemd-main.rs
@@ -1,0 +1,8 @@
+mod systemd;
+pub mod common;
+
+use clap::Parser;
+
+fn main() {
+    systemd::Cli::parse().call(module_path!())
+}
